+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - 2020-01-24
+## 0.1.2 - 2020-01-27
+
+### Fixed
+- fix generator description bug not catched by tests
+
+## 0.1.1 - 2020-01-27
+
+### Added
+- import directive support
+- generator always renews all generated _twirp and _pb files from all protos to support import directive
+- lib/twirp added to $LOAD_PATH
+
+## 0.1.0 - 2020-01-24
 
 ### Added
 - mount_twirp route helper